Fix integer dtype for empty per-period observable-index arrays
iys_real (per-measurement-row observable indices) defaulted to
float64 when a period has zero real dynamic measurements (an empty
list comprehension), since np.array([]) infers float64 without an
explicit dtype. _get_iy_trafos now gathers via trafo_by_index[iys],
which requires integer indices, so this surfaced as "IndexError:
arrays used as indices must be of integer (or boolean) type" for
models with such periods (e.g. Blasi_CellSystems2016 in the
benchmark collection).

+++ python/sdist/amici/sim/jax/petab.py
@@ -659,7 +659,8 @@ class JAXProblem(eqx.Module):
                         [
                             self.model.observable_ids.index(oid)
                             for oid in m[petabv2.C.OBSERVABLE_ID].values
-                        ]
+                        ],
+                        dtype=int,
                     )
                     iy_trafos_real = _get_iy_trafos(
                         iys_real,
